[LOOP CRC32]Add Crc32 Optimization in Gzip For crc32 algorithm in APBC int_gzip. Match crc32 lookup table algorithm. An example for crc32 lookup table alg: ```c do { c = crc_32_tab[((int)c ^ (*s++)) & 0xff] ^ (c >> 8); } while (--n);
Usage: `gcc -O3 -march=armv8.1-a -floop-crc yourfile.c`
Node: The cpu you use needs to support the crc32 instructions

+++ gcc/testsuite/gcc.dg/tree-ssa/loop-crc-sucess.c
@@ -73,12 +73,11 @@ ulg updcrc(s, n)
         c = 0xffffffffL;
     } else {
         c = crc;
-        if (n) 
-        do {
-            c = crc_32_tab[(c ^ (*s++)) & 0xff] ^ (c >> 8);
+        if (n) do {
+            c = crc_32_tab[((int)c ^ (*s++)) & 0xff] ^ (c >> 8);
         } while (--n);
     }
     crc = c;
     return c ^ 0xffffffffL;       /* (instead of ~c for 64-bit machines) */
 }
-/* { dg-final { scan-tree-dump-times "the loop can be optimized" 1 "loop_crc"} } */
+/* { dg-final { scan-tree-dump-times "The 1th loop form is successmatched,and the loop can be optimized." 1 "loop_crc"} } */
